+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ---
 1.1.4.8v124 - 20260311:
-- upgrade dependencies: ligit2 to 1.9.2, openssl to 3.6.1
+- upgrade dependencies: libgit2 to 1.9.2, openssl to 3.6.1
 - optimize english translation (pr #125, thx @clach04)
 - optimize code of `PullToRefreshBox` (pr #128, thx @msmt2018)
 - align c libs to 16KB page size
@@ -11,7 +11,7 @@
 
 
 
-- 升级依赖: ligit2 to 1.9.2, openssl to 3.6.1
+- 升级依赖: libgit2 to 1.9.2, openssl to 3.6.1
 - 优化英语翻译 (pr #125, thx @clach04)
 - 优化 `PullToRefreshBox` 的代码 (pr #128, thx @msmt2018)
 - 对齐c库为16KB页大小